cancels, sim stuff

+++ nerago-gearing/main/main.go
@@ -9,6 +9,7 @@ import (
 	"paladin_gearing_go/items"
 	"paladin_gearing_go/model"
 	"paladin_gearing_go/setup"
+	"paladin_gearing_go/simulate"
 	"paladin_gearing_go/util"
 	"runtime/pprof"
 	"time"
@@ -42,7 +43,9 @@ func main() {
 		defer func() {
 			pprof.StopCPUProfile()
 			f.Close()
-			os.Rename(f.Name(), files.ProfileDir+"main.pgo")
+			if simulate.WowSimRanDuringCurrentProcess {
+				os.Rename(f.Name(), files.ProfileDir+"main.pgo")
+			}
 		}()
 	}
 
@@ -89,11 +92,11 @@ func core(printer *util.PrintRecorder) {
 	// statWeightsGridIntoRanking(printer)
 	// statWeightsCustom(printer)
 
-	statWeights_CompareAlgorithms(printer)
+	//statWeights_CompareAlgorithms(printer)
 
 	// statWeightsGrid_updateAll(printer)
 
-	// PaladinMultiRun(printer)
+	PaladinMultiRun(printer)
 
 	// findUpgrades_Paladin(printer)
 }
